Several fixes for wxPython 3.0.2.0
These are the backwards-compatible fixes from the 15-wx3-upgrade branch
that did not introduce any complex behavior changes (e.g. no pngcrush).

Fix for crashing Colors dialog

Same problem as in 360a516.
Need to check if this is wx3-only or if it's needed in Phoenix as well.
Also need to think about wrapping ComboBox in the wx-begone branch to
fix this in a non-hacky way, no GUI component should ever hard-crash the
application if it's not used correctly :P

Fix mod checker crashing when closed and reopened

Was caused by Link.Frame.modChecker never being unset - there was even a
TODO for it in the code.

Fix focus not being lost on TextCtrls

This is precisely the reason why my new event framework Skip()'s by
default. Not skipping by default is incredibly toxic and makes it really
easy to create incredibly strange behavior.

Enough ranting - the problem was that several text fields weren't losing
focus anymore. Easiest to reproduce with the ones on the Mods tab. Made
editing them almost impossible since wx had to idea what to do with the
cursor when multiple text fields were active at the same time.

Delay progress dialog when unpacking to project

Under wx3, it would shove itself over any warning messages / rename
dialogs we showed. This wasn't really a *problem*, since the user could
just move the progress dialog out of the way, but it was annoying. This
is a fairly painless fix that just asks the questions first and then
creates the progress dialog.

Clean up some outdated wx usages

Minor step towards Phoenix compatibility.

Co-authored-by: MrD <the.ubik@gmail.com>