Upgrade many Python 2-only ur'' strings to Python 2/3 concatenations
In Python 3, ur'' is a syntax error. But since concatenating a 'raw'
string with a Unicode string in Python 2 gives another Unicode string,
we can simply concatenate lengthy raw regular expressions with an empty
Unicode string instead, which doesn't change the Python 2 behaviour and
simply becomes a no-op in Python 3.