FO3, FNV: Fully decode PERK
For some reason we had this in mergeClasses, but the definition was
nowhere near done. It was missing the script variables section, and
EPFD did not map fids (which will lead to a disaster if we merge a
mod... hopefully people didn't break their installations *too* hard).

This fixes all issues I found with the FO3/FNV PERK definition and makes
them match xEdit almost completely (see the TODO in there).