Disable empty sublist removal by default
Adds an option to reenable it to the leveled list patcher. See the
relevant issue below for further discussion on why this is can cause
issues and should be disabled by default, only to be enabled by users
who know what they're doing(TM). Left it enabled by default for Oblivion
though, since that's the one game where it might be somewhat useful
(plus years of history of people expecting it to be there...).

Closes #475