FO3, FNV: Don't upgrade IMGS/DNAM format
If we upgrade this to the newer format, then interior lighting for some
merged mods breaks. If you don't believe me, grab Clarity
(https://www.nexusmods.com/newvegas/mods/62481/?), merge it into the BP
and start a new game of FNV. Note how dark all the interiors are? This
commit forces WB to write out the data exactly how it read it in as
well. Would obviously break patching, but we don't patch IMGS since it's
purely a graphics record.