Drop MelFull0
One of the reasons for starting this branch was to exterminate this ugly
hack with extreme prejudice. It's only needed for 5 types in Oblivion,
but has to be checked for *every* subrecord for every record type in all
games. Yuck. _MelDistributor was the final nail in the coffin that
allowed this to be dropped.