TES5: Rewrite VMAD
This is a revived and finished version of the (very) old lojack-vmad
branch. Revived this because our PERK implementation is broken without
it (whoops). It's also another significant leap towards
patching/merging/ESL-checking every record type.

Almost completely rewritten to update it for changes, PEP8 it, heavily
reduce code duplication, (hopefully) make refactoring and moving it to
brec eventually possible (for FO4 and future games), etc.

Co-authored-by: Infernio <infernio@icloud.com>