Allowing merging masters if their children are merged
This is something CBash alrady does, so let's just extend the logic to
PBash as well. Tested it and it works fine.
Basically, say you have mod A.esp, which is a master of B.esp and C.esp.
A.esp is technically mergeable, but merging it would deactivate B's and
C's master, and is therefore impossible. However, if we can merge both
B.esp and C.esp into the BP, then we can also merge A.esp, since it's
now only a master of the BP itself.

Co-authored-by: MrD <the.ubik@gmail.com>