AMultitweakItem refactoring - tweak_name and tweak_tip:
Went ahead and renamed name -> tweak_name just to realize that it is
actually unused. Actually those name/tip belong to gui but it's ok to be
here, otherwise we need a gui hierarchy for tweaks and that is an
overkill.
Bonus: got rid of yet another 'label' attribute - label is already used
in wx/balt
Some 'tweak_read_classes' leaked from future 312-patchers merge - I left
them here it doesn't harm.
I replaced double quoted strrings with single quoted ones.
Note we would like to drop the __init__overrides - for instance
OppositeGenderAnimsPatcher and CBash_VORB_NPCSkeletonPatcher, the latter
uses a cache that will be recalculated each time for PBash but not for
CBash leading to subtle bugs.
Another WIP part is further reworking init - for instance add the
p_file in - however that's part of #312 refactoring which changes
how those are instantiated anyway.
Yet another WIp is tweak tips are actually a bit broken, some display
"set ... to:" but that's cosmetics. Could be solved by turning tweak_tip
to a property.
Finally note FIXME in ANamesTweak_BodyTags - this patcher must be
reworked to actually do its thing once and as soon as the patch is
launched.
Merging despite all this cause 312 merge depends on it and actually
already simplifies quite a few things