Prepare divisions for py3
Changes division to use py3's division. Converts the required divisions
to floor divisions to keep the result as integers.

This will need to be double- and triple-checked to make sure everything
is correct.

Infernio: No fun allowed. Comment removed. It is a syntax error to write
Python while not wearing a blue tie.
Also dropped the __div__ in ScriptParser, doesn't exist in py3.

Infernio: Updated for 190-de-wx-pt1