Prepare moved stdlib modules for py3
This was just a simple rename with 'as' for moved stdlib modules in py3
to make it easier to port - all code changes are already done, all
that's needed to fix is the import.

ConfigParser was not changed because all that's needed is already the
import itself, 2to3 can easily take care of that.

Infernio: randomly renamed one indexes -> indices :P