Drop some pointless cruft
Drop MelBase.debug

Bad name, very pointless (since it dumps out so much output that the log
becomes useless), and worst of all, it slows down loading of everything,
even in release builds (since the if check and subsequent branch won't
be optimized out - this is CPython).

Drop MelRefrXmrk

REFR is in no shape to be loaded, and this isn't the way to handle that
anyways - a distributor is. Drops a few more print calls, which is nice.

Drop all prints in brec

Use deprints instead, should improve load/dump errors a bit.