binary.py
Was writing utility classes for PEX reading/writing and it ocurred to me
that I was just rewriting the ones I made for VMAD - so time to move all
that complexity into an abstract module that the rest of WB can use.
Could add it to bolt, but that's enough of a kitchen sink already. We
could even move some parts of bolt in here (e.g. bolt.Flags).