WIP: Plug in the Scripts tab
Creates the GUI for it. Commented out, same as the BSAs tab. We badly
need to wrap TreeCtrl to provide a better API for it - in particular,
this is *extremely* slow for 'large' PEX files (i.e. 20KB+. Yup, it's
*that* bad) and would benefit from a virtual API that uses callbacks
to construct parts of the tree only when they're needed.

This also functions as a trial run for the 'Data Tab' (# 50), which
I'd like to use a TreeCtrl for as well. That will *necessitate* such
a virtual API - if construction of a tree for a PEX file is this slow
with wxPython, imagine what a full Data folder looks like...