FO3/FNV: Merge LVL* implementation back into brec
Had to add a new class variable to accomplish this, which should
probably be updated for Skyrim/FO4 as well. This also brings all fixes
and improvements (e.g. the 255+ items fix) that have landed in brec's
implementation to FO3 and FNV. Needed for the patchers refactoring, in
which we unify the ListsMerger and FidListsMerger classes.