Plug in the new LOOT parser
Initial test run of this parser show the following results:
 - It successfully parses the v0.15 masterlists for all games
 - It successfully parses and evaluates all tag conditions in the
   v0.15 branches of the masterlists for all games
 - It's *very* fast - may actually be slightly faster than
   libloot-python? If it is, then it's probably due to us now avoiding
   a second parse of every plugin (syscalls, record decoding, etc.)

Note that by far the slowest operation is parsing the YAML. This is
handled by PyYAML's C-based parser, which is only available if LibYAML
is installed. Thankfully, all Windows wheels come with it statically
linked, but Linux users may have to install a separate package to get
access to the C-based parser. If the parser is not available for
whatever reason, we fall back to a pure Python-based one. Much slower,
but still not a death sentence (~4.5s vs ~0.6s on my machine).