Properly handle game-specific xEdit and CK names
I *was* just going to introduce a quick bush.game variable and use it in
the approriate strings, but one thing led to another, the CS/CK got
involved, and it turned into a bigger refactoring that also features a
bunch of typo corrections, improved translation possibilities, etc.

But the main draw is that we now always show the right name for xEdit
and the CS/CK/GECK, unless I missed something :P

Note that I added GameInfo.xe as a class for this - I expect it to grow
with further refactoring in the future, as mentioned in links.py.
Another important part of this is GameInfo.cs -> GameInfo.ck. This
isn't due to my bias against the CS, but due to bolt.Path.cs already
existing and thereby making any searches for '.cs' annoying.

Closes #239