Gracefully handle reverting to corrupt files
Offer the user an option to abort the revert procedure, restoring the
original file again. Randomly stumbled across this due to the FIXME (at
least I'm assuming this is what the FIXME referred to?) for snapshots,
but this also implements it for backups.

Note that I'm keeping the temp backup around - gets cleaned up at exit
anyways, and it may give computer-savvy users an out if they
accidentally click 'No' on the new popup.

Also no longer swallows the traceback for the FileError, deprinting it
to the BugDump instead so that one can actually figure out why the
reverting failed.