Misc cleanups, better patch log l10n capabilities
Instead of _(u'whatever') + u': %d', always use _(u'whatever: %d'). Not
only is it shorter and easier to understand at a glance, but it also
makes it possible to translate this for cultures that place their
numbers elsewhere - or even ones that use completely non-Western number
formatting.

Dropped importers changes, most were to CBash importers (which aren't
long for this world) and others should be centralised at some point
anyways (they're all log changes).