[!] Encapsulate long/short FormID handling
Doesn't fully get rid of getLongMapper yet, since the tweaks will be
tackled in the rest of this branch (duh).

Required a bit of rewriting here and there, but the resulting code is
actually much nicer.

This is a serious behavior change. I was not able to get it to break,
despite testing it fairly extensively on all games, but that doesn't
mean anything - hence the [!].