save_headers: Reimplement SSE headers with lz4
Works nicely, mirrors how short and succinct the zlib version is -
pretty happy with this :)

Note that we gain no speedup from this, because we now read the entire
file instead of just the (very short) header table. However, this does
of course open the door to support editing the save headers - see next
few commits.