Optimize GameDecider a bit
Mopy/bash/env.py:
scripts/utils.py:
Don't use _ as iteration variable

Clashes *badly* with the global _ for translation, causing errors upon
usage.

Mopy/bash/bosh/__init__.py:
Better error message for missing strings

Actually mention the names we looked for and the BSAs we considered.

Mopy/bash/brec.py:
Mopy/bash/parsers.py:
Tiny refactoring for readability

Under #480

Mopy/bash/_games_lo.py:
SSE: Make a copy of plugins.txt for the CK

The CK looks for plugins.txt in the game folder, not AppData\Local,
which is where it *should* be looking. So make a copy of the file
whenever we write it out.

Note that this is fixed in FO4, and seems to crash Skyrim LE's CK when a
plugins.txt file is present, so we only do it for SSE.