bosh: dds_files.py
Pre-alpha at best. DDS is one of mankind's worst inventions, and I hope
it will be years before I have to touch it again.

Tried to keep this API general enough to allow expansion for full
reading/writing in the future, but it's good enough for DX10 BA2
extraction for now.