BashFrame(WindowFrame):
- yey
- note most of naked uses of _native_widget are in ask*** calls, maybe
a methods mixin? Tricky part is Link.Frame - or maybe parent=None?
- hiding the buttons is broken fixed on next commit by Infernio.

Note:

@@ -100,2 +100,4 @@ def _resolve(obj):
             return obj
+        elif obj is None:
+            return None
         else:

Mopy/bash/gui/top_level_windows.py:

@@ -75,3 +75,8 @@ def set_icons(self, wx_icon_bundle):

-    def close_win(self): self._native_widget.Close()
+    def close_win(self, force_close=False):

Maybe move Destroy inside close win? or not?
docs from wx docstring - just this once (or twice)

Remaining traceback on closing:

Traceback (most recent call last):
  File "bash\gui\events.py", line 167, in _post
    result = listener(*listener_args)
  File "bash\balt.py", line 1567, in _conversation_wrapper
    return func(*args, **kwargs)
  File "bash\basher\__init__.py", line 3831, in RefreshData
    BashFrame.saveListRefresh(focus_list=False)
  File "bash\basher\__init__.py", line 4027, in saveListRefresh
    BashFrame.saveList.RefreshUI(focus_list=focus_list)
  File "bash\balt.py", line 1787, in RefreshUI
    self.PopulateItems()
  File "bash\balt.py", line 1768, in PopulateItems
    item = self.GetItem(index)
  File "bash\balt.py", line 2078, in GetItem
    return self.__gList.FindItemAt(index)
  File "bash\balt.py", line 1549, in FindItemAt
    return self._itemId_item[self.GetItemData(index)]
KeyError: 0

Turns out this is related to the double binding of refresh - so probably
refresh runs a second time and by then the control is empty? Before IIUC
the second binding was ignored - we must add guards in our API (done
later).

EDIT: Actually passing evt_active instead of nothing:

@@ -3807 +3807 @@ def set_status_info(self, infoTxt):
-    def RefreshData(self, event=None, booting=False):
+    def RefreshData(self, evt_active=True, booting=False):
@@ -3811 +3811 @@ def RefreshData(self, event=None, booting=False):
-        if event and not event.GetActive() or self.inRefreshData: return
+        if not evt_active or self.inRefreshData: return

@@ -108 +108,2 @@ def __init__(self, parent, title, icon_bundle=None, _base_key=None,
-        self.on_activate = EventHandler(self._native_widget, _wx.EVT_ACTIVATE)
+        self.on_activate = EventHandler(self._native_widget, _wx.EVT_ACTIVATE,
+            arg_processor=lambda event: [event.GetActive()])

also fixes the error. So probably was called when frame was deactivated.
Anyway needs some work - we really want this to run exactly once and
exactly when we regain focus while we were previously hidden, so *not*
when a dialog *we* display is shown.

Note: return EventResult.FINISH, this is also a fix for above and
mimics previous behavior (for better or worse).

Infernio reverted stuff like

@@ -238,5 +237,4 @@ def OnImport(self):
         except Exception as e:
-            balt.showError(_AComponent._resolve(Link.Frame), _(
-                u'An error occurred reading from ') + inPath.stail +
-                           u':\n\n%s' % e)
+            balt.showError(Link.Frame, _(u'An error occurred reading from ')
+                           + inPath.stail + u':\n\n%s' % e)
         self.UpdateUIButtons()

as he moved all this to the base methods