Upgrade to wxPython 4
After the wx3 upgrade (see #15), the next step towards Python 3 is
moving to wx4, which is the first wxPython version to support py3.

Note that I dropped it from the installer - that's broken, but it's
already broken on dev and nightly since it doesn't install all our
other dependencies like:
 - chardet
 - lz4
 - pyyaml
I don't see any way to update it do to these either, we'd have to
hardcode direct links to PyPI resources, and I doubt PyPA would like
that :)

So I propose we just drop the ability of the installer to install python
versions. Seeing as not a single person has complained about this being
broken since we moved to pypi chardet, I doubt anyone uses it - plus
would simplify the UI a bit (no more checkmarks). Done in the next
commit.

wx4: Build fixups

Yay, it's a mess :/
Note I dropped MPR.dll, has been in our nightlies for a long time now
but clearly *shouldn't* be there, it's a standard Windows DLL for
communicating with network drivers.

Closes #488
Under #460

Co-authored-by: MrD <the.ubik@gmail.com>
Co-authored-by: lojack5 <1458329+lojack5@users.noreply.github.com>