More fixups and wrapping
Note us dropping a whole bunch of explicit indices handling :)

Installer_Espm_DeselectAll.Execute is scary - but seems to be more
correct than before(!) - the previous implementation placed strings with
asterisks into espmNots if a plugin was renamed. Something in BAIN
handled this correctly and removed the asterisk, but the new
implementation is *much* simpler and doesn't do that.