Prune ClosePanel and ShowPanel overrides:
I had initially hoped that I could use EVT_SHOW to completely detangle
the GUI part from the rest of (data model and UI) operations ShowPanel
does - turns out as noted in the comments of SashUIListPanel.ShowPanel
that EVT_SHOW is too early. We do want to reduce occurrences of those,
they are one of the trickiest code paths - related to data refreshing.
That's it for now apart from renaming to populate_panel?