WithFirstShow(_AComponent):
Tried using it in parent as in NotebookPanel(WithFirstShow, PanelWin),
but did not quite work -> NotebookPanels must be fully populated for
this to work which is not the case with various variations - from
Installers not at all to Mods/Saves that the details sashes are not
set correctly. Plus let's avoid event handling when we can call
functions manually.