Rework installer - drop python installation
Broken since 8e201c49bbc809da89b1bda1d269f4cb7619dfc0, plus not
possible to repair it now that we're switching to wx4.

This drops the Standalone/Python version checkboxes in favor of simply
always installing the standalone version if the game's checkbox is
checked. Simplifies the installer a bit as well, which is nice.

I fixed a few bugs along the way, like the uninstall pages having
untranslated strings, and also made it unnecessary to edit macros.nsh
whenever we rename or delete a python file. Also dropped the AII part of
the installer, since it was Oblivion-specific and unnecessary - on first
launch, WB will do that by itself, and once we drop the templates in
favor of writing them using bsa_files, this will have to go anyways.

Under #460, #488