Regression: Fix TESCS/GECK not launching with xSE
Broke during be92f4beb1028dc12199fb61dc805087a28d0c17. Not exactly a
surprise, this code was (and still is) perilous waters.

Problem with this button is that we have to switch executables
completely - the regular CK executable with no arguments for newer
games or if the xSE toggle is off, but the xSE executable with the
'-editor' argument for older games iff the xSE executable is present and
the xSE toggle is activated.