[!]Move ini_files to top:
Original motivation was to use in parsing the game ini as soon as
possible, but this needs more work (rewrite the ConfigParser code).
Turns out this is a move in the right (init management) direction.
Will be useful in morrowind and in general - reworked to get
rid of a very nasty coupling between ini_files and bosh globals
(oblivionIni and iniInfos).

- Fixup in target_ini_exists logic (note return msg)
- Decouple ini_files from balt

Tagged an [!] as I now instantiate *but* the main game ini as GameIni:

@@ -3221,5 +3217,5 @@ def initBosh(bashIni, game_ini_path):
     gameInis = [oblivionIni]
-    gameInis.extend(OblivionIni(dirs['saveBase'].join(x), 'cp1252') for x in
+    gameInis.extend(IniFile(dirs['saveBase'].join(x), 'cp1252') for x in
                     bush.game.iniFiles[1:])

I followed uses and those should (must) not be GameIni instances
but I may have missed smth.

Signed-off-by: MrD <the.ubik@gmail.com>