Drop INIPanel.ShowPanel refresh_target param
Seems to be True in almost all cases - notable exception is when we
switch to the tab from a different tab, which is why it breaks when
editing the load order in Morrowind. This feels like a sledgehammer
solution though :/

Utumno: untested load_order function to keep previous behavior,
other solutions would be too verbose/complicated. At least we
switched the default :)