Drop Undelete Refs and a bunch of cruft
This is all stuff that xEdit is really good at, and that we're really
terrible at. So terrible that PBash was happily writing out the most
corrupt plugins I have *ever* seen when you tried to use it.

The remainder is either unused cruft or has been disabled for a long
time now.

Under #173