localize: Fixup for dumping new languages
When I tightened this except in b3faf0173c42cb3299a52db5181638f1b6eed169,
I forgot to put in IOError, so dumping for new languages 'broke' (as in,
it definitely isn't as nice as it could be - right now it just fails to
find an old file and throws an IOError, which is then deprinted).

Mopy/bash/bosh/__init__.py:
Deprint when we initialize each TableFileInfos

Very noisy, but allows us to tell when in the init procedure something
blew up without having to look at exact line numbers. Also made it log
the folders it's trying to create, helpful when diagnosing that someone
moved their game folders around.