gui.Table
Very cool and powerful component for displaying tabular data. Based on
wx.grid.Grid, but with some useful enhancements (that took a *long* time
to program) like Ctrl+A, Ctrl+C, Ctrl+V, and an overall easier to use
API.

Note that this is currently only built around the use case I have for it
in gui_fomod, so general purpose methods for using it don't really exist
yet. I'm putting them off so we can decide on a design when/if we
actually need it.