Remove backwards compat for 306 from BAIN: EEE
            self.__setstate(values)
drop print - values are unicode if basestring

If someone wants to downgrade from this commit it should be 307 at least

- self.vdata['boltPaths'] is never checked
- the crc_installer dict is not used

Then I went ahead and removed the Installers internal dicts backwards
compat. This slowed down shutdown and load - bear in mind that for
big packages these structures could have thousands or even tens of
thousands of keys and GPath is not the fastest function around. Loading
was equally slowed down but I had already added isinstance checks to
checks for `LowerDict` - I was flirting with the idea of only pickling
std classes (due to may years of hassles of settings compat and the
bitter Path experience), but exactly because of this load/save overhead
I abandoned this idea. After all we want a LowerDict here and it won't
go away anytime soon.

Apart from that those are the core BAIN structures and every occurrence
counts - see # 219 <---  RRR for the gory details. Note my tricks for permitting
upgrades to Py3 (get(unicode) or get(bytes)). Some prefixes added.

vdata[u'version'] = 2 is WIP - by the time we get to py3 we should add
backwards compat code to PickleDict also - would it make sense to keep
this bytes (no)? Plus if this is gonna be any use we should check for
it and abort if settings are a newer version (think those LowerDict
checks - I could not use vdata dict) - so we must when the dust settles
add a version check in PickleDict so older versions would warn or abort
if we try loading newer settings - or at least backup.

Absorbed some __getattribute__ pruning

Closes EEE (more stuff needed?) # 376 <---  RRR