Move Wizard Images out of top-level dir
You probably didn't even know that wizards could specify 'Wizard
Images\\No.jpg' and get the No.jpg image from the Wizard Images folder -
without having to include it in their own package. This was of course
intended to eventually become a library of common images for wizards,
but that never panned out.

Having that folder exposed to the user in the Mopy directory has always
been ugly IMO, so now it sits in the Mopy/bash/images folder. And
instead of checking relative to the Mopy folder, we now check relative
to this folder for fallbacks. That way we preserve backwards
compatibility *and* make it much easier to use any of the 600+ images we
bundle for wizards (you could technically do this previously by
specifying e.g. 'bash\\images\\wryesplash.png', but this was
undocumented and so not actually an intended feature - now it's
documented and you can just do 'wryesplash.png').

 Mopy/bash/bosh/bsa_files.py:
Don't blow up on corrupt BSA headers

Raise a BSAError instead, which will be caught and will allow us to
handle the corrupt BSA properly.