Split Invent into three tags
All this started when someone on Discord noticed that Import Inventory
does not actually import changes, only additions and removals. However,
its documentation clearly stated that it would patch 'additions, changes
and removals'. So I took the opportunity to add that to PBash - but to
preserve backwards compatibility, I obviously couldn't change the
behavior of Invent.

So I decided to split the tag into three tags (Invent.Add, Invent.Change
and Invent.Remove) and introduce an alias that interprets the old Invent
tag as Invent.Add + Invent.Remove whenever it finds it in descriptions,
LOOT tags and BashTags files. I also updated CBash to use the new tags,
but didn't bother to try and make it understand Invent.Change. So CBash
behaves exactly like before. One ugly wart here is the local imports in
mods_metadata.py, but that file might as well be _mods_metadata.py for
all the rethinking it deserves :P

This introduces a framework to allow us to easily support indefinite
backwards-compatibility aliases for Bash Tags, so I went ahead and
refactored the old InventOnly and C.GridFlags tag to use this system.
Also documented all removed and deprecated tags in the Advanced Readme.

An eventual goal here is to investigate if we can refactor this new
implementation of Import Inventory to create an abstract _AMerger, then
rename _SimpleImporter to _APreserver and split importers.py into
_cbash_importer.py, mergers.py and preservers.py.

Note the addition of MelObject.__hash__, noticed way too late that my
set lookups wouldn't work. It's a bit hacky, but at least now any other
set usages of MelObject that might be hiding in the patchers won't break
anymore. Also dropped the weird 'removable' logic in there, was broken
at first but after fixing it it became a noop.

Under #312