Fix fomod case sensitivity issue
If the case differed between the ModuleConfig and the actual files on
disk, they wouldn't get installed. To fix this, it turned out to be
necessary to invert the entire fomod dict. It's a LowerDict, so using
'x in values()' does *not* respect case. Plus the new implementation is
much easier to understand and definitely faster.

Unfortunately, this necessitated backwards compat code which we'll have
to drop in 308.

Mopy/bash/basher/gui_fomod.py:
Fix Select All selecting unselectable options