Offer a 'smart' install command for newbies
This does what most mod managers do when you ask them to install
something: show a wizard if it exists, and only fall back to manual
configuration if that's not possible. Can be enhanced in the future when
we let users pick a Data folder to use from the package to also handle
unrecognized archives by popping such a dialog up.

Implementation is not the nicest (having to call _initData, mostly), but
the upside is that it works perfectly without actually duplicating any
of the code from those links.

Obsoletes the need for #108 as well.