cosaves: Support DATA chunks
Added by SOS to Skyrim LE/SE. No, not Sounds of Skyrim, the other one.
Need to support this since it contains a list of mod names that have to
be remapped when we change a save master.

Features the worst way to store integers I've ever seen, namely by
converting them to strings and then writing the literal ASCII bytes of
the string out. Who doesn't love needing O(log n) space to store an int?