Move record_groups to brec
Not sure about this yet - I like it because it groups all of the
low-level PBash code together in one package. But it might be bloating
brec *too* much?

Note the imports in patch_files were unused, and me having to make
tests/__init__.py emulate the startup a bit closer so that bush has _()
available.