Allow only one game master per game
This variable (bush.game.masterFiles) was created for the half-baked
Nehrim support we're dropping on this branch. Skyrim and FO4VR used it
incorrectly, it's meant to be used for alternative master files that can
completely replace the original one.

Follow-up might be to drop bosh.modInfos.masterName? Not sure that's a
good idea though.