Absorb ImportFactions into _APreserver
No idea why this was using any of that nonsense, it's just a dead simple
'preserves' patcher and can therefore use the 100% standard
_APreserver pattern. It looks like it may have been based off of the
Import Relations patcher, and was intended to merge additions and
deletions, but during development its author decided that was a bad
idea and made it a standard preserver. In fact, that seems to be what
the CBash patcher does. We should weight our options here in the future
and either make this a merger or keep it as a preserver. Either one is
much easier now that it's absorbed :)

Leaked from future parser-abc merge.