Absorb NamesPatcher into _APreserver
The CELL/WLRD stuff in there was unused (CELL and WRLD were not listed
in any game's namesTypes). Which is good, because it didn't work.

I would like to get it working and add CELL/WRLD to namesTypes (as well
as readding DIAL, which I had to remove from Skyrim), plus this is a
good step in the direction of absorbing CellImporter.