Absorb NpcFacePatcher into _APreserver
Note that using NpcFaces + one of the two 'Only' tags was undefined
behavior, which changes with this commit. Thankfully there aren't any
mods on the LOOT masterlist that have this mistake.

Finally figured out what those 'fid attrs' that also appear in the
Graphics patcher are for. They are supposed to denote any FormID
attributes that we should verify before importing data from that record
into the BP. Namely, if the FormID points to a missing master, we should
skip the record from being imported at all.
The ones used for the Graphics patcher are woefully outdated. Also
probably worth adding this for other patchers?

Also had to add support for the ForceFullImport tag to _APreserver.
Only two preservers left :)