cosaves: Catch struct errors when loading
So that we can provide a better error message (now points out the
problematic cosave). Also went ahead and created some custom exceptions
for cosaves.

We definitely need to rethink some core FileInfo stuff here - right now
it will show the corresponding save as having an error when it's
actually the *cosave* that has the error. That's part of exposing the
cosaves better in the UI though, which is blocked on allowing trees
instead of just UILists (ref #529).

Mopy/bash/basher/__init__.py:
Changed the save warning since it will fire for much more than just
unknown header versions.

Under #336