TES5: Add PERK\DESC and ACTI\RNAM to Text
PERK was commented out because nothing else was patching it, which was
because it was an entirely unfinished definition. ACTI wasn't included
simply because I missed it on my first pass over this tag.

Took this opportunity to prefix text_types; also prefixed the
defaultEnabled strings because they just happened to catch my eye while
scrolling :P