Get rid of some MelUnion mapping duplication
The idea is to resolve tuples keys into their elements for the final
mapping, so that something like

MelUnion({
    0: foo,
    1: foo,
    2: foo,
    3: bar,
    4: bar,
    5: bar
}, decider=...)

can be replaced with

MelUnion({
    (0, 1, 2): foo,
    (3, 4, 5): bar,
}, decider=...)

Doesn't look like much there, but some of the 'foo's and 'bar's can be
several lines long and quite complex, so not duplicating them is a very
good idea.

MelLocation is a good example of how useful this can be. Also took the
opportunity to prefix all unions.