Fix for merging interior CELLs into the BP
Would blow up on the next mod that has a WLRD otherwise since it got
added to LoadFactory.topTypes. Still not sure what the point of that one
is, might be ready to go in favor of just extracting the core of
addClass into a new method and having that call itself with
CELL/DIAL/WRLD when one of the child record types is encountered.

Should have been part of the future fixup merge earlier - whoops.

Under #480