Support 'unknown is unused' behavior for flags
Stolen from xEdit, allows us to drop that awful hack in CellImporter.
Also grepped through xEdit's sources and added the kwarg to all flags
that have it in xEdit's definitions - with some exceptions to common
subrecords that will be refactored in the next few commits.

Also includes some fixups for BODT loading (the legacy flags are an u8
with three unused garbage bytes afterwards, not a u32).