Properly implement BSA load order
We had two attempts to discern the right load order strewn about the
codebase: one in ModInfo._extra_bsas, and another in
ModInfos.get_active_bsas. This commit centralizes it in
ModInfos.get_bsa_lo and also adds support for the Skyrim VR setting
'sVrResourceArchiveList', which overrides all plugin BSAs.

As a result of this commit, BAIN will now show conflicts with the base
game's BSAs as well, a long-standing TODO. Note that this might make the
Conflicts tab really hard to read for mods that override a lot of
vanilla BSA assets (e.g. USSEP). Which really just demonstrates that the
text-based tabs won't do in the long run and should be replaced (at
least optionally) with proper GUI components (i.e. a tree).

Was also able to drop the vanilla_string_bsas - means we no longer need
to figure those out and set them for each new game. To keep startup just
as fast as it was before, I introduced a heuristic to avoid having to
scan a bunch of unnecessary BSAs.

Note: Contains a fixup(?) for BSAs loaded from INIs. If two INIs had the
same resource key, we'd add the BSAs twice. Working on the assumption
that the first INI we find it in wins right now, but we may have to
merge them instead? More testing needed.

Mopy/bash/bosh/bain.py:
Some minor refactoring on find_conflicts - should be entirely
standalone, i.e. not depend on any external state (bass.settings).

Closes #534
Closes #546