Small refactoring for DefaultIniFile - more do_update:
- original idea of adding a stat_file=True param to AFile.init, that
would be False for DefaultIniFile. Turns out it's enough to override
_stat_tuple, as the behavior is then confined inside the anomalous case
of DefaultIniFile, and the rest of the AFile API works out of the box.
We also considered a VirtualFile - but if in doubt...
Nice thing is we finally can call super.__init__, which is a requirement
for a good API (and a PyCharm warning).

Note that if raise_on_error == True, do_update() will _reset_cache(null)
AFile.__init__() won't - but then it's _null_stat already
Mopy/bash/ini_files.py: rename __empty -> __empty_settings

Infernio:
- Mopy/bash/ini_files.py: some prefixing for the bytes part

Under #336

Co-authored-by: Infernio <infernio@icloud.com>