Make 'text replacer' tweaks static
Step 1 to making these much more powerful and extendable. Also happens
to make them far more readable, since you no longer have to parse a
soup of parentheses and commas.

I've officially declared war on DynamicTweak. At the very least it
should be reserved for settings tweaks, but I think we could save quite
a bit of duplication (if not LOC) and make the code much more
approachable for newcomers if we dropped it even there.