TES4: Add Marksman -> Archery tweak
Requested on Discord, turned out to be hell.
A simple re.sub is not powerful to handle it. So I could write a method
and pass it to re.sub, but that quickly turned into ~20 lines of ugly
duplicated code.

So I turned my attention to the rule of thumb that has never failed me:

  In case of doubt, build a tool in bolt.

And so I present: esubs. The name might not be fantastic, but two hard
things and all that. An esub callable can be created via
bolt.build_esub by passing it a string which will then be parsed.
The syntax is identical to re.sub's, with the extension of the '$d(s)'
construct. This construct resolves to 's' with the case changed to match
that of the 'd' group.

That probably seems like the most niche thing you've ever heard of. An
example might help:

Consider this regex: \b(a|A) (m|M)arksman\b
This will match 'a marksman', 'a Marksman', 'A marksman' and 'A
Marksman'. We want to replace 'marksman' with 'archer', so we write
something like:

  re.sub(r'\b(a|A) (m|M)arksman\b', r'\1n \2rcher', u'a marksman')

and get 'an mrcher' back. Right, because \2 will resolve to 'm' or 'M',
not the 'a' or 'A' we want. And that's where esubs can help:

  re.sub(r'\b(a|A) (m|M)arksman\b', build_esub(r'\1n $2(a)rcher'),
         u'a marksman')

and if you enter this into a repl (pasting in build_esub first), you
will indeed see that it returns 'an archer'.

But the weirdest part of this whole thing is that I managed to beat
the stdlib's performance:

  100000 loops, best of 3: 2.9 usec per loop
  1000000 loops, best of 3: 1.77 usec per loop

I gave them both a '\1rcher' string. Yeah, I beat the stdlib in pure
Python. No idea how this happened. Difference is less drastic on py3,
but esubs are still faster (1.8 usec vs 1.0 usec). Only thing I could
possibly think of would be the stdlib having to parse the string JIT,
whereas I parse it AOT.