[!] Drop CBash
Time to sunlight this. It has been effectively unmaintained since 2011,
when Waruddar stopped updating it, and completely unmaintained since
2017, when leandor last worked on it. Some more reasons to drop it, in
no particular order:

- The way it hooks into WB (cint.py) is poor. ctypes is a terrible way
  to write such a complicated and involved wrapper.
- It is buggy to hell and back. Some highlights are Import Relations not
  doing the thing it's supposed to do (preserving imported data) and the
  NoMerge tag completely breaking importing.
- It means we can't really rename record attributes if they also occur
  in cint.py, since we can't change any of those.
- It's a binary that we have to keep in our repo. Ugh.
- The way the patchers are hooked into WB is poor too due to way too
  much duplication between patchers. Just look at how much code is
  gone now (687 insertions(+), 21977 deletions(-))
- It stands in the way of py3 due to its monstrous and arcane nature -
  no one is really sure what encodings it needs, which strings should be
  unicode and which shouldn't, etc.
- The implementation in CBash_PatchFile was completely arcane. To this
  day, I have no clue what it was doing. So much getattr/hasattr
  hackery to substitute for what probably should have been handled with
  inheritance.

We released 307 with CBash deprecated (#520) precisely to prepare for
this, giving users one stable release worth of deprecation time. Trying
to use a CBash config now will just give an error message and tell the
user to convert it in 307.

Tested in all games, BPs build fine. Insta-merging because this will
conflict *hard* with all other branches, and I'm not looking to repeat
tweak pooling :P

Under #312, #460 and a bunch more issues
Closes #530