[!]WIP getNumRecords:
Do we want to include the groups in Recalculate Record Counts? Turns
out those are not saved in the GRUP header so does not matter (RRC is
a debug menu) so kept them. Only diff (on CoblMain) is the CELL group

mod_links.py  120 Execute: CELL GRUP has 159 records

vs

mod_links.py  120 Execute: CELL GRUP has 178 records

When the group is loaded in getNumRecords its GRUP header is already
loaded (enforce?) so the code was correct in saying `+ includeGroups`
still added all other group headers separately (and avoided the seek
for groups taking size for the ride).

This will change BP record counts however so !

WIP in all generality (see rest of the branch) and in particular this is
essentially Mob* area (aka last frontier, bleeding edge, #480, and as
your manager puts it great fun). Note to self:

https://en.uesp.net/wiki/Skyrim_Mod:Mod_File_Format#Groups

This is a core API and we want to retain generality - meaning we don't
want an abc MobBase - getNumRecords is a use case for a mod groups API
and we want to be fast (not load complex stuff). Mapping the usecases
should be step number one

Mopy/bash/brec/mod_io.py: hack for info label - ignore I solved this
at the root in debug_strs commit (next)

Mopy/bash/brec/record_groups.py: added includeGroups param [!]