Backwards-compatible changes from 2to3 run
Plus manual changes in the vicinity to follow up on the 2to3 changes
(e.g. some prefixing, sorted([...]) -> sorted(...), etc.). Note the
negative line diff :)

Took this opportunity to finish what 2to3 started and replace all
set(x for x in ...) / dict((x, y) for x, y in ...) with the appropriate
literals. Much faster (up to a 40% improvement on py3) and more
readable (IMO).

Note I changed a couple usages that were like this:

    my_dict = {x: f(x) for x in my_list}
    my_list.sort(key=my_dict.__getitem__)

to

    my_list.sort(key=lambda x: f(x))

It's possible the former is faster, though I can't imagine it would be
(has to run over the whole list twice for a start).

Mopy/bash/barg.py:
Took this opportunity to prefix _short_to_long and make editing it less
miserable in the future.

requirements.txt:
Bump toml requirement to get pycharm to shut up

Under #460