Implement the revamped New Plugin dialog
Allows choosing any number of masters from the current load order, as
well as specifying the extension and flags.

Also includes a search bar, because I wanted to finally justify having
added that class :)

And because it wasn't over-engineered enough, I also made it impossible
to create plugins with >255 masters by blocking the OK button and
showing a warning in the GUI (would otherwise have blown up on dump).