Future-proof master/espm/esl limits
Probably not likely that we'll get a game that allows > 255 masters, but
let's nip this duplication in the bud before anyone has to go hunting
through the codebase if I end up eating my words.

The LO changes are because the max_plugins method was never used as a
whole, only ever one of its two return values. So splitting it into two
methods makes it much more readable:

    sort_offset = load_order.max_plugins()[0] - 1

vs.

    sort_offset = load_order.max_espms() - 1