Initial high DPI awareness
Basically just means marking ourselves as aware on Windows and only
assigning size/position after creating the widget so that we can treat
all our pixel specifications as device-independent pixels (DIP).

However, that caused the game select popup to become misaligned and ugly
- because it uses raw wx, of course. This turned into a wild goose chase
and ended up with me rewriting all wx usages in bash.py and de-wx'ing a
lot of it.

Overall, this is probably the best showcase of what #190 can do for us.
This would have been flat out *impossible* without it. We would have
had to sprinkle FromDIP/csf calls all over the codebase and play
whack-a-mole with every dialog and every popup. Instead, we got away
with some ~200 lines of refactoring in bash.py and a handful of
FromDIP/csf calls in gui <3

Remaining work is creating high-resolution versions of all the icons we
use in our GUI, of course.

Under #190 for de-wx'ing of bash.py
Closes #555 - but as mentioned above, over the course of #190 we'll want
to create SVG versions of all our icons so we can get crisp,
high-resolution images in our GUI without having to include tons of
different sizes (ref #557 and #511).